refactor app.py (5/7): extract scratch helpers to helpers/scratch.py
load_scratch_work, save_scratch_work move to helpers/scratch.py.
SCRATCH_WORK_FILE constant redefined locally.

Smoke-tested: GET /scratch/work and POST /scratch/work with JSON
body — both 200 (load + save round-trip).
